fixed optimization
before: I was checking if n was even, then changing step to 2. I also didn't actually have step included in the for loop, so it only made the time marginally slower. 
after: I should have been checking if n was odd to change step to 2, since odd numbers cannot have even divisors, while even numbers can have odd and even divisors. Also added the step into the for loop, so now the 25% speed optimization actually is implemented (not that it really makes much of a difference on this computation)